extract an interface, clean up
