+++ quesma/quesma/router.go
@@ -34,7 +34,7 @@ import (
 	"time"
 )
 
-func configureRouter(cfg *config.QuesmaConfiguration, sr schema.Registry, lm *clickhouse.LogManager, ip *ingest.IngestProcessor, console *ui.QuesmaManagementConsole, phoneHomeAgent telemetry.PhoneHomeAgent, queryRunner *QueryRunner, indexRegistry *index_registry.IndexRegistry) *mux.PathRouter {
+func configureRouter(cfg *config.QuesmaConfiguration, sr schema.Registry, lm *clickhouse.LogManager, ip *ingest.IngestProcessor, console *ui.QuesmaManagementConsole, phoneHomeAgent telemetry.PhoneHomeAgent, queryRunner *QueryRunner, indexRegistry index_registry.IndexRegistry) *mux.PathRouter {
 
 	// some syntactic sugar
 	method := mux.IsHTTPMethod
