+++ quesma/quesma/ui/management_console.go
@@ -101,14 +101,14 @@ type (
 		phoneHomeAgent            telemetry.PhoneHomeAgent
 		schemasProvider           SchemasProvider
 		totalUnsupportedQueries   int
-		indexRegistry             *index_registry.IndexRegistry
+		indexRegistry             index_registry.IndexRegistry
 	}
 	SchemasProvider interface {
 		AllSchemas() map[schema.TableName]schema.Schema
 	}
 )
 
-func NewQuesmaManagementConsole(cfg *config.QuesmaConfiguration, logManager *clickhouse.LogManager, indexManager elasticsearch.IndexManagement, logChan <-chan logger.LogWithLevel, phoneHomeAgent telemetry.PhoneHomeAgent, schemasProvider SchemasProvider, indexRegistry *index_registry.IndexRegistry) *QuesmaManagementConsole {
+func NewQuesmaManagementConsole(cfg *config.QuesmaConfiguration, logManager *clickhouse.LogManager, indexManager elasticsearch.IndexManagement, logChan <-chan logger.LogWithLevel, phoneHomeAgent telemetry.PhoneHomeAgent, schemasProvider SchemasProvider, indexRegistry index_registry.IndexRegistry) *QuesmaManagementConsole {
 	return &QuesmaManagementConsole{
 		queryDebugPrimarySource:   make(chan *QueryDebugPrimarySource, 10),
 		queryDebugSecondarySource: make(chan *QueryDebugSecondarySource, 10),
